app/requirements.txt
-i https://pypi.org/simple
attrs==20.3.0
automat==20.2.0
bytesbufio==1.0.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
cycler==0.10.0
flask==1.1.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.0
kiwisolver==1.3.1
lxml==4.6.2 ; platform_python_implementation == 'CPython'
markupsafe==1.1.1
matplotlib==3.3.4
nltk==3.5
numpy==1.20.0
oauthlib==3.1.0
parsel==1.6.0
pillow==8.1.0
protego==0.1.16
pyasn1-modules==0.2.8
pyasn1==0.4.8
pybase64==1.1.4
pycparser==2.20
pydispatcher==2.0.5
pyhamcrest==2.0.2
pyopenssl==20.0.1
pyparsing==2.4.7
pysocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
queuelib==1.5.0
regex==2020.11.13
requests-oauthlib==1.3.0
requests[socks]==2.25.1
scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
textblob==0.15.3
tqdm==4.56.0
tweepy==3.10.0
twisted==20.3.0
urllib3==1.26.3
w3lib==1.22.0
werkzeug==1.0.1
wordcloud==1.8.1
zope.interface==5.2.0
